requirements.txt
django==1.9.2
djangorestframework
